:sparkles: Feat: 공연 상세보기 구현
- 공연 상세보기 API 구현

+++ backend/src/main/java/com/capsule/youkids/festival/repository/FestivalRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface FestivalRepository extends JpaRepository<Festival, Integer> {
+public interface FestivalRepository extends JpaRepository<Festival, Long> {
     @Query("select new com.capsule.youkids.festival.dto.view.FestivalRecommItemDto(f)"
             + "from Festival f where f.festivalChildId in :festivalIds")
     List<FestivalRecommItemDto> getMixedRecommfestivals(@Param("festivalIds") List<Long> festivalIds);
